fix database connection error

+++ src/main/java/com/contactmanager/app/domain/ContactRepository.java
@@ -10,6 +10,6 @@ public interface ContactRepository extends CrudRepository<Contact, Long> {
 	List<Contact> findByEmail(String email);
 	List<Contact> findByPhone(int phone);
 	List<Contact> findById(int id);
-	List<Contact> deleteById(int id);
+	List<Contact> deleteById(long id);
 }
